ccmgr: Adapted and improved CC management scripts
Now using the Python Virtual Environment layer

+++ tools/ccmgr/cckill.sh
@@ -1,32 +1,41 @@
 #!/bin/bash
 
-PID_FILE="run/cc.pid"
+DBDIR="$PWD/db"
+PID_FILE="$DBDIR/cc.pid"
+
 KILLED="no"
+
+if [ ! -d "$DBDIR" ]; then
+    echo "$PWD is not a CC directory"
+    exit 1
+fi
 
 if [ -f "$PID_FILE" ]; then
     CC_PID=`cat "$PID_FILE"`
     
-	if [ -d "/proc/$CC_PID" ]; then
+    if [ -d "/proc/$CC_PID" ]; then
         # Shut down Control Center.
         echo "Shutting down CC..."
         kill -s SIGTERM "$CC_PID"
         for count in $(seq 1 10)
         do
+            if [ ! -d "/proc/$CC_PID" ]; then
+                break
+            fi
             sleep 1
-            test -d "/proc/$CC_PID" || break
         done
         if [ -d "/proc/$CC_PID" ]; then
             echo "Graceful shutdown failed, now killing CC."
             kill -s SIGKILL "$CC_PID"
         fi
-		KILLED="yes"
-	else
-		echo "Cleaned stale pid file ($CC_PID)"
-	fi
+            KILLED="yes"
+    else
+        echo "Cleaned stale pid file ($CC_PID)"
+    fi
 fi
 
 if [ "$KILLED" == "yes" ]; then
-	echo "CC shut down ($CC_PID)"
+    echo "CC shut down ($CC_PID)"
 else
-	echo "CC was not running"
+    echo "CC was not running"
 fi
